+++ src/entities/cloudconnection.cpp
@@ -212,8 +212,8 @@ bool CloudConnection::fillFromQuery(const QSqlQuery &query) {
     this->name = query.value(QStringLiteral("name")).toString();
     this->serverUrl = query.value(QStringLiteral("server_url")).toString();
     this->username = query.value(QStringLiteral("username")).toString();
-    this->password = CryptoService::instance()->decryptToString(
-        query.value(QStringLiteral("password")).toString());
+    QString storedPassword = query.value(QStringLiteral("password")).toString();
+    this->password = CryptoService::instance()->decryptToString(storedPassword);
     this->priority = query.value(QStringLiteral("priority")).toInt();
 
     const int databaseVersion =
@@ -274,7 +274,11 @@ bool CloudConnection::store() {
     query.bindValue(QStringLiteral(":username"), this->username);
     query.bindValue(QStringLiteral(":account_id"), this->accountId);
     query.bindValue(QStringLiteral(":password"),
-                    CryptoService::instance()->encryptToString(this->password));
+                    this->id > 0
+                        ? CryptoService::instance()->encryptToString(
+                              this->password,
+                              QStringLiteral("database/cloudConnection/%1/password").arg(this->id))
+                        : CryptoService::instance()->encryptToString(this->password));
     query.bindValue(QStringLiteral(":priority"), this->priority);
     query.bindValue(QStringLiteral(":qownnotesapi_enabled"), this->appQOwnNotesAPIEnabled);
 
@@ -316,7 +320,11 @@ bool CloudConnection::storeMigratedCloudConnection() {
     query.bindValue(QStringLiteral(":serverUrl"), this->serverUrl);
     query.bindValue(QStringLiteral(":username"), this->username);
     query.bindValue(QStringLiteral(":password"),
-                    CryptoService::instance()->encryptToString(this->password));
+                    this->id > 0
+                        ? CryptoService::instance()->encryptToString(
+                              this->password,
+                              QStringLiteral("database/cloudConnection/%1/password").arg(this->id))
+                        : CryptoService::instance()->encryptToString(this->password));
     query.bindValue(QStringLiteral(":priority"), this->priority);
 
     if (!query.exec()) {